requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0.9
numpy>=1.24
pillow>=10.0
scikit-learn>=1.3
umap-learn>=0.5.6
matplotlib>=3.7
streamlit>=1.36
plotly
